requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
flask-limiter
marshmallow
pytest
gunicorn
python-dotenv
flasgger
werkzeug
faker
pytest-cov
black
flake8